Finish the rest of documentation off, restructure files
